Allow people to search almost anywhere
Results can't be searched yet, but in practice, this should be fine. The
issue with searching results is that results are stored outside the
database, and can't be easily searched. We would have to store results
in the database for that, and I'm not too keen on it. Maybe as another
iteration...